+++ src/editor/components/header/TheOptionsAction.vue
@@ -84,9 +84,15 @@ export default Vue.extend({
     background: none !important;
     color: #000 !important;
     border: none !important;
-    font-size: 16px !important;
-    line-height: 18px !important;
+    font-size: 15.5px !important;
     padding: 0 !important;
+    line-height: 20.5px !important;
+
+    .b-icon.bi {
+      display: inline-block !important;
+      overflow: visible !important;
+      vertical-align: -0.15em !important;
+    }
 
     &::after {
       display: none !important;